implem. annual investment
