+++ src/app/app.module.ts
@@ -3,10 +3,11 @@ import { AppComponent } from './app.component';
 import { HeaderComponent } from './header/header.component';
 import { UserInputComponent } from './user-input/user-input.component';
 import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [AppComponent, HeaderComponent, UserInputComponent],
   bootstrap: [AppComponent],
-  imports: [BrowserModule],
+  imports: [BrowserModule, FormsModule],
 })
 export class AppModule {}
