+++ src/app/user-input/user-input.component.ts
@@ -5,4 +5,10 @@ import { Component } from '@angular/core';
   templateUrl: './user-input.component.html',
   styleUrl: './user-input.component.css',
 })
-export class UserInputComponent {}
+export class UserInputComponent {
+  enteredInitialInvestment = '0';
+  enteredAnnualInvestment = '0';
+  enteredExpectedReturn = '5';
+  enteredDuration = '10';
+  onSubmit() {}
+}
